+++ resolwe/elastic/tests/test_index.py
@@ -220,6 +220,7 @@ class IndexTest(ElasticSearchTestCase):
         user_1 = user_model.objects.create(username='user_one')
         user_2 = user_model.objects.create(username='user_two')
         user_3 = user_model.objects.create(username='user_three')
+        user_pub = user_model.objects.get(username='public')
         group = Group.objects.create(name='group')
 
         # Create test object
@@ -233,20 +234,53 @@ class IndexTest(ElasticSearchTestCase):
         self.assertEqual(es_objects[0].groups_with_permissions, [group.pk])
         self.assertEqual(es_objects[0].public_permission, False)
 
-        # Change permissions
-        remove_perm('view_testmodel', user_2, test_obj)
+        # Add user permission
         assign_perm('view_testmodel', user_3, test_obj)
+
+        es_objects = TestSearchDocument.search().execute()
+        self.assertEqual(es_objects[0].users_with_permissions, [user_1.pk, user_2.pk, user_3.pk])
+        self.assertEqual(es_objects[0].groups_with_permissions, [group.pk])
+        self.assertEqual(es_objects[0].public_permission, False)
+
+        # Remove user permission
+        remove_perm('view_testmodel', user_2, test_obj)
 
         es_objects = TestSearchDocument.search().execute()
         self.assertEqual(es_objects[0].users_with_permissions, [user_1.pk, user_3.pk])
         self.assertEqual(es_objects[0].groups_with_permissions, [group.pk])
         self.assertEqual(es_objects[0].public_permission, False)
 
-        # Change permissions
+        # Remove group permission
+        remove_perm('view_testmodel', group, test_obj)
+
+        es_objects = TestSearchDocument.search().execute()
+        self.assertEqual(es_objects[0].users_with_permissions, [user_1.pk, user_3.pk])
+        self.assertEqual(es_objects[0].groups_with_permissions, [])
+        self.assertEqual(es_objects[0].public_permission, False)
+
+        # Add group permission
+        assign_perm('view_testmodel', group, test_obj)
+
+        es_objects = TestSearchDocument.search().execute()
+        self.assertEqual(es_objects[0].users_with_permissions, [user_1.pk, user_3.pk])
+        self.assertEqual(es_objects[0].groups_with_permissions, [group.pk])
+        self.assertEqual(es_objects[0].public_permission, False)
+
+        # Add public permission
         assign_perm('view_testmodel', AnonymousUser(), test_obj)
 
         es_objects = TestSearchDocument.search().execute()
+        self.assertEqual(es_objects[0].users_with_permissions, [user_pub.pk, user_1.pk, user_3.pk])
+        self.assertEqual(es_objects[0].groups_with_permissions, [group.pk])
         self.assertEqual(es_objects[0].public_permission, True)
+
+        # Remove public permission
+        remove_perm('view_testmodel', AnonymousUser(), test_obj)
+
+        es_objects = TestSearchDocument.search().execute()
+        self.assertEqual(es_objects[0].users_with_permissions, [user_1.pk, user_3.pk])
+        self.assertEqual(es_objects[0].groups_with_permissions, [group.pk])
+        self.assertEqual(es_objects[0].public_permission, False)
 
     def test_field_name(self):
         from .test_app.models import TestModel