+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
 Fixed
 -----
 - Make manager more robust to ORM/database failures during data object processing
+- Rebuild the ElasticSearch index after permission is removed from an object
 
 
 ==================